requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
dj-database-url==0.5.0
Django==3.0
django-filter==2.2.0
gunicorn==20.0.4
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
vine==1.3.0
whitenoise==5.0.1
zipp==0.6.0
psycopg2==2.8.4